Move approximation caching code to `edge` module
