+++ crates/fj-kernel/src/algorithms/approx/shell.rs
@@ -4,7 +4,7 @@ use std::collections::BTreeSet;
 
 use crate::objects::Shell;
 
-use super::{curve::CurveCache, face::FaceApprox, Approx, Tolerance};
+use super::{edge::CurveCache, face::FaceApprox, Approx, Tolerance};
 
 impl Approx for &Shell {
     type Approximation = BTreeSet<FaceApprox>;
