+++ crates/fj-kernel/src/algorithms/approx/sketch.rs
@@ -4,7 +4,7 @@ use std::collections::BTreeSet;
 
 use crate::objects::Sketch;
 
-use super::{curve::CurveCache, face::FaceApprox, Approx, Tolerance};
+use super::{edge::CurveCache, face::FaceApprox, Approx, Tolerance};
 
 impl Approx for &Sketch {
     type Approximation = BTreeSet<FaceApprox>;
